Create jellyfin group in proxmox role
I have decided to use traditional UNIX permissions to control access
to the NFS shares. For this to work, I will need the user/groups
with access to be present on both nfs server and client. This will
be best done in the future with a SSO like freeIPA, but for now the
users/groups will be created with Ansible